Moved check_date to general utils Better print for testing api
