+++ practical_porcupines/discord_bot/bot.py
@@ -1,12 +1,8 @@
 import discord
 from discord.ext import commands
-from practical_porcupines.utils import ConfigBot
+from practical_porcupines.utils import ConfigBot, check_date
 from practical_porcupines.discord_bot.api import get_difference
-from practical_porcupines.discord_bot.utils import (
-    embed_generator,
-    decode_diff_resp,
-    check_date,
-)
+from practical_porcupines.discord_bot.utils import embed_generator, decode_diff_resp
 
 config_bot = ConfigBot()
 
@@ -20,4 +16,6 @@ async def on_ready():
     """
 
     print(f"{bot_client.user.name} is online with the id of '{bot_client.user.id}'!")
-    print(await decode_diff_resp(await get_difference("2010", "2019")))  # API testing
+    print(
+        "Testing api" + await decode_diff_resp(await get_difference("2010", "2019"))
+    )
