+++ practical_porcupines/discord_bot/utils.py
@@ -34,25 +34,3 @@ async def decode_diff_resp(difference_obj):
         return decoded_obj["body"]["wl_difference"]  # Return difference part
     except:
         return "ERROR 1001: API returning wrong values!"
-
-
-def check_date(date):
-    """
-    > Gets date
-    - String: Date
-    < Returns matched date
-    x Returns None
-    """
-
-    date_pattern = (
-        "((19[0-9]{2}|2[0-9]{3})((:((0[1-9]|1[012]):"  # year & month
-        "([123]0|[012][1-9]|31))(:([01][0-9]|2[0-3]):"  # day and hour
-        "([0-5][0-9]):([0-5][0-9]))?)?)?)"  # minute and second
-    )
-
-    date_match = re.findall(date_pattern, date)
-
-    if not date_match:
-        return None
-
-    return date_match[0][0]
